fix changes of jose
- some changes were lost
- it is necessary to put more attention while doing merges
- keep your work/folders organized
- it was the best intent to fix the files
