+++ src/App.js
@@ -1,23 +1,19 @@
-import React from "react";
-import "./App.css";
+import React from 'react';
+import './App.css';
 
-import TaskForm from "./TaskForm";
-import TaskAdd from "./Task";
-import TaskList from "./TaskList";
-
-import taskForm from "./taskForm";
+import TaskItem from './components/TaskItem';
+import TaskList from './components/TaskList';
 
 function App() {
   return (
-    <div className="todoApp">
-      <div className="top-botton"></div>
+    <div className='todoApp'>
+      <div className='top-botton'></div>
 
-      <div className="task-list-main">
+      <div className='task-list-main'>
         <h1> My to-do </h1>
 
-        <TaskForm />
-
-        <taskForm />
+        <TaskItem />
+        <TaskList />
       </div>
     </div>
   );
