+++ src/components/TaskForm/TaskForm.css
@@ -1,0 +1,29 @@
+.taskForm {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: center;
+}
+
+.task-input {
+  width: 350px;
+  font-size: 18px;
+  background-color: white;
+  padding: 14px 32px 14px 16px;
+  border-radius: 4px 0 0 4px;
+  border: 2px solid #00471b;
+  outline: none;
+}
+
+.task-button {
+  padding: 16px;
+  font-size: 18px;
+  font-family: sans-serif;
+  border: none;
+  border-radius: 0 4px 4px 0;
+  cursor: pointer;
+  outline: none;
+  background-color: #00471b;
+  color: #fff;
+  text-transform: capitalize;
+}
