+++ src/components/TaskForm/index.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import './TaskForm.css';
+
+function TaskForm(props) {
+  return (
+    <form className='taskForm'>
+      <input
+        className='task-input'
+        type='text'
+        placeholder='Add new task '
+        name='text'
+      />
+      <button className='Task-button'>Add Task</button>
+    </form>
+  );
+}
+export default TaskForm;
